Dzung commit valide input  + delete 1 to many

+++ StudentManagerGroup1/src/main/java/com/studentmanager/controller/SinhVienController.java
@@ -5,10 +5,12 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import com.studentmanager.entity.SinhVien;
+import com.studentmanager.service.DiemRepository;
 import com.studentmanager.service.SinhVienRepository;
 
 @Controller
@@ -17,12 +19,14 @@ public class SinhVienController {
 	// Mapping table SinhVien
 	@Autowired
 	private SinhVienRepository sinhVienRepository;
+	@Autowired
+	DiemRepository diemRepository;
+
 	@RequestMapping("/sinhVienList")
 	public String listCustomer(Model model) {
 		model.addAttribute("listSinhVien", sinhVienRepository.findAll());
 		return "TableStudentResult";
 	}
-
 
 	@RequestMapping("/sinhVienUpdate/{maSV}")
 	public String updateSinhVien(@PathVariable String maSV, Model model) {
@@ -30,6 +34,7 @@ public class SinhVienController {
 		if (sinhVien.isPresent()) {
 			model.addAttribute("selectedGender", sinhVien.get().isGioiTinh());
 			model.addAttribute("sinhVien", sinhVien.get());
+			System.out.println(sinhVien.get());
 		}
 		return "SinhVienUpdate";
 	}
@@ -40,6 +45,7 @@ public class SinhVienController {
 		model.addAttribute("listSinhVien", sinhVienRepository.findAll());
 		return "redirect:/sinhVienList";
 	}
+
 	@RequestMapping("/insertInfoSinhVien")
 	public String redirectSinhVien() {
 		return "SinhVienInsert";
@@ -55,9 +61,15 @@ public class SinhVienController {
 
 	@RequestMapping("/sinhVienDelete/{maSV}")
 	public String doDeleteSinhVien(@PathVariable String maSV, Model model) {
-		sinhVienRepository.deleteById(maSV);
-		model.addAttribute("listSinhVien", sinhVienRepository.findAll());
-		return "redirect:/sinhVienList";
+		if (diemRepository.findByMaSV(maSV).size() == 0) {
+			sinhVienRepository.deleteById(maSV);
+			model.addAttribute("listSinhVien", sinhVienRepository.findAll());
+			return "redirect:/sinhVienList";
+		} else {
+			model.addAttribute("deletedMsg", "Bạn cần xoá thông tin trong bảng Điểm của sinh viên có mã SV = " + maSV +" trước!!!");
+			model.addAttribute("listSinhVien", sinhVienRepository.findAll());
+			return "TableStudentResult";
+		}
 	}
 
 }
